+++ topobank/manager/models.py
@@ -12,9 +12,16 @@ class Surface(models.Model):
 
     There can be many topographies (measurements) for one surface.
     """
+    CATEGORY_CHOICES = [
+        ('exp', 'Experimental data'),
+        ('sim', 'Simulated data'),
+        ('dum', 'Dummy data')
+    ]
+
     name = models.CharField(max_length=80)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     description = models.TextField(blank=True)
+    category = models.TextField(choices=CATEGORY_CHOICES, null=True, blank=False)
 
     def num_topographies(self):
         return self.topography_set.count()