+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2
 
+- added category to surfaces (experimental/simulated/dummy data, #145)
 - show units to users even if they cannot be changed (#149)
 - in surface summary, sort bandwidth bars by lower bound (#85)
 - added statistics on welcome page (#148)